# api/swagger.yaml
openapi: 3.0.1
info:
  title: Emerald Web Api
  version: v1.0.0
servers:
- url: /
security:
- Bearer: []
paths:
...
  /api/Admin/checkListCategory/create:
    post:
      tags:
      - Admin
      operationId: apiAdminCheckListCategoryCreatePOST
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/CheckListCategoryCreateRequest'
            example:
              Name: Shakir
              ParentId: 45
          application/json:
            schema:
              $ref: '#/components/schemas/CheckListCategoryCreateRequest'
            example:
              Name: Shakir
              ParentId: 45
          text/json:
            schema:
              $ref: '#/components/schemas/CheckListCategoryCreateRequest'
            example:
              Name: Shakir
              ParentId: 45
          application/*+json:
            schema:
              $ref: '#/components/schemas/CheckListCategoryCreateRequest'
            example:
              Name: Shakir
              ParentId: 45
      responses:
        200:
          description: Success
      x-swagger-router-controller: Admin
  /api/Admin/note/category:
    post:
      tags:
      - Admin
      operationId: apiAdminNoteCategoryPOST
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/ApplicationNoteCategoryRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationNoteCategoryRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/ApplicationNoteCategoryRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ApplicationNoteCategoryRequest'
      responses:
        200:
          description: Success
      x-swagger-router-controller: Admin
...
  /api/Admin/guarantorRoles/create:
    post:
      tags:
      - Admin
      operationId: apiAdminGuarantorRolesCreatePOST
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/GuarantorRolesCreateRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/GuarantorRolesCreateRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/GuarantorRolesCreateRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/GuarantorRolesCreateRequest'
      responses:
        200:
          description: Success
      x-swagger-router-controller: Admin
...
  /api/Admin/maritalStatus/create:
    post:
      tags:
      - Admin
      operationId: apiAdminMaritalStatusCreatePOST
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/MaritalStatusCreateRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/MaritalStatusCreateRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/MaritalStatusCreateRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/MaritalStatusCreateRequest'
      responses:
        200:
          description: Success
      x-swagger-router-controller: Admin
  /api/Admin/businessType/create:
    post:
      tags:
      - Admin
      operationId: apiAdminBusinessTypeCreatePOST
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/BusinessTypeCreateRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessTypeCreateRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/BusinessTypeCreateRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/BusinessTypeCreateRequest'
      responses:
        200:
          description: Success
      x-swagger-router-controller: Admin
  /api/Admin/loanProducts/create:
    post:
      tags:
      - Admin
      operationId: apiAdminLoanProductsCreatePOST
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/LoanProductCreateRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/LoanProductCreateRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/LoanProductCreateRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/LoanProductCreateRequest'
      responses:
        200:
          description: Success
      x-swagger-router-controller: Admin
  /api/Admin/RolePermissions/{RoleId}:
    get:
      tags:
      - Admin
      operationId: apiAdminRolePermissionsRoleIdGET
      parameters:
      - name: RoleId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          default: ""
      responses:
        200:
          description: Success
      x-swagger-router-controller: Admin
  /api/Admin/RolePermissions:
    post:
      tags:
      - Admin
      operationId: apiAdminRolePermissionsPOST
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/RolePermissionModificationRequest'
            example:
              RolePermissions:
              - RolePermissionId: 2
                PermissionId: 1
                CanEdit: false
                CanView: false
                CanAdd: false
              - RolePermissionId: 0
                PermissionId: 2
                CanEdit: false
                CanView: false
                CanAdd: true
              - RolePermissionId: 0
                PermissionId: 3
                CanEdit: false
                CanView: true
                CanAdd: true
              RoleId: 04145692-b5c9-411c-b33d-671261a2ebe0
          application/json:
            schema:
              $ref: '#/components/schemas/RolePermissionModificationRequest'
            example:
              RolePermissions:
              - RolePermissionId: 2
                PermissionId: 1
                CanEdit: false
                CanView: false
                CanAdd: false
              - RolePermissionId: 0
                PermissionId: 2
                CanEdit: false
                CanView: false
                CanAdd: true
              - RolePermissionId: 0
                PermissionId: 3
                CanEdit: false
                CanView: true
                CanAdd: true
              RoleId: 04145692-b5c9-411c-b33d-671261a2ebe0
          text/json:
            schema:
              $ref: '#/components/schemas/RolePermissionModificationRequest'
            example:
              RolePermissions:
              - RolePermissionId: 2
                PermissionId: 1
                CanEdit: false
                CanView: false
                CanAdd: false
              - RolePermissionId: 0
                PermissionId: 2
                CanEdit: false
                CanView: false
                CanAdd: true
              - RolePermissionId: 0
                PermissionId: 3
                CanEdit: false
                CanView: true
                CanAdd: true
              RoleId: 04145692-b5c9-411c-b33d-671261a2ebe0
          application/*+json:
            schema:
              $ref: '#/components/schemas/RolePermissionModificationRequest'
            example:
              RolePermissions:
              - RolePermissionId: 2
                PermissionId: 1
                CanEdit: false
                CanView: false
                CanAdd: false
              - RolePermissionId: 0
                PermissionId: 2
                CanEdit: false
                CanView: false
                CanAdd: true
              - RolePermissionId: 0
                PermissionId: 3
                CanEdit: false
                CanView: true
                CanAdd: true
              RoleId: 04145692-b5c9-411c-b33d-671261a2ebe0
      responses:
        200:
          description: Success
      x-swagger-router-controller: Admin
  /api/Admin/user/Create:
    post:
      tags:
      - Admin
      operationId: apiAdminUserCreatePOST
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/ApplicationUserRequest'
            example:
              Id: "1"
              FirstName: adeel
              LastName: Hassan
              Email: ABS
              UserName: AVChgshsshs
              PasswordHash: absafasfc
              RoleName: User
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationUserRequest'
            example:
              Id: "1"
              FirstName: adeel
              LastName: Hassan
              Email: ABS
              UserName: AVChgshsshs
              PasswordHash: absafasfc
              RoleName: User
          text/json:
            schema:
              $ref: '#/components/schemas/ApplicationUserRequest'
            example:
              Id: "1"
              FirstName: adeel
              LastName: Hassan
              Email: ABS
              UserName: AVChgshsshs
              PasswordHash: absafasfc
              RoleName: User
          application/*+json:
            schema:
              $ref: '#/components/schemas/ApplicationUserRequest'
            example:
              Id: "1"
              FirstName: adeel
              LastName: Hassan
              Email: ABS
              UserName: AVChgshsshs
              PasswordHash: absafasfc
              RoleName: User
      responses:
        200:
          description: Success
      x-swagger-router-controller: Admin
  /api/Admin/user/update:
    post:
      tags:
      - Admin
      operationId: apiAdminUserUpdatePOST
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/ApplicationUserUpdateRequest'
            example:
              Id: "1"
              FirstName: Adeel
              LastName: hassan
              Email: adeelhassan592@gmail.com
              UserName: asd
              PasswordHash: asf
              RoleName: asf
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationUserUpdateRequest'
            example:
              Id: "1"
              FirstName: Adeel
              LastName: hassan
              Email: adeelhassan592@gmail.com
              UserName: asd
              PasswordHash: asf
              RoleName: asf
          text/json:
            schema:
              $ref: '#/components/schemas/ApplicationUserUpdateRequest'
            example:
              Id: "1"
              FirstName: Adeel
              LastName: hassan
              Email: adeelhassan592@gmail.com
              UserName: asd
              PasswordHash: asf
              RoleName: asf
          application/*+json:
            schema:
              $ref: '#/components/schemas/ApplicationUserUpdateRequest'
            example:
              Id: "1"
              FirstName: Adeel
              LastName: hassan
              Email: adeelhassan592@gmail.com
              UserName: asd
              PasswordHash: asf
              RoleName: asf
      responses:
        200:
          description: Success
      x-swagger-router-controller: Admin
  /api/Admin/Role/Create:
    post:
      tags:
      - Admin
      operationId: apiAdminRoleCreatePOST
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/ApplicationRoleRequest'
            example:
              Name: adeel
              NormalizedName: adeel
              ConcurrencyStamp: ad
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationRoleRequest'
            example:
              Name: adeel
              NormalizedName: adeel
              ConcurrencyStamp: ad
          text/json:
            schema:
              $ref: '#/components/schemas/ApplicationRoleRequest'
            example:
              Name: adeel
              NormalizedName: adeel
              ConcurrencyStamp: ad
          application/*+json:
            schema:
              $ref: '#/components/schemas/ApplicationRoleRequest'
            example:
              Name: adeel
              NormalizedName: adeel
              ConcurrencyStamp: ad
      responses:
        200:
          description: Success
      x-swagger-router-controller: Admin
  /api/Admin/role/update:
    post:
      tags:
      - Admin
      operationId: apiAdminRoleUpdatePOST
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/ApplicationRoleRequest'
            example:
              Name: adeel
              NormalizedName: adeel
              ConcurrencyStamp: ad
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationRoleRequest'
            example:
              Name: adeel
              NormalizedName: adeel
              ConcurrencyStamp: ad
          text/json:
            schema:
              $ref: '#/components/schemas/ApplicationRoleRequest'
            example:
              Name: adeel
              NormalizedName: adeel
              ConcurrencyStamp: ad
          application/*+json:
            schema:
              $ref: '#/components/schemas/ApplicationRoleRequest'
            example:
              Name: adeel
              NormalizedName: adeel
              ConcurrencyStamp: ad
      responses:
        200:
          description: Success
      x-swagger-router-controller: Admin
...
components:
  schemas:
    CheckListCategoryCreateRequest:
      required:
      - name
      type: object
      properties:
        name:
          maxLength: 100
          type: string
        parentId:
          type: integer
          format: int32
      additionalProperties: false
    ApplicationNoteCategoryRequest:
      type: object
      properties:
        categoryName:
          type: string
      additionalProperties: false
    GuarantorRolesCreateRequest:
      required:
      - role
      type: object
      properties:
        role:
          maxLength: 50
          type: string
      additionalProperties: false
    MaritalStatusCreateRequest:
      required:
      - status
      type: object
      properties:
        status:
          maxLength: 50
          type: string
      additionalProperties: false
    BusinessTypeCreateRequest:
      required:
      - business
      type: object
      properties:
        business:
          maxLength: 50
          type: string
      additionalProperties: false
    LoanProductCreateRequest:
      required:
      - productName
      type: object
      properties:
        productName:
          maxLength: 50
          type: string
        perc:
          type: number
          format: double
          nullable: true
      additionalProperties: false
    RolePermissionViewModel:
      type: object
      properties:
        rolePermissionId:
          type: integer
          format: int32
        permissionId:
          type: integer
          format: int32
        permissionName:
          type: string
        canEdit:
          type: boolean
        canView:
          type: boolean
        canAdd:
          type: boolean
      additionalProperties: false
    RolePermissionModificationRequest:
      type: object
      properties:
        rolePermissions:
          type: array
          items:
            $ref: '#/components/schemas/RolePermissionViewModel'
        roleId:
          type: string
      additionalProperties: false
    ApplicationUserRequest:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        userName:
          type: string
        passwordHash:
          type: string
        roleName:
          type: string
      additionalProperties: false
    ApplicationUserUpdateRequest:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        userName:
          type: string
        passwordHash:
          type: string
        roleName:
          type: string
      additionalProperties: false
    ApplicationRoleRequest:
      type: object
      properties:
        name:
          type: string
        normalizedName:
          type: string
        concurrencyStamp:
          type: string
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: apiKey
      description: "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter\
        \ 'Bearer' [space] and then your token in the text input below.\r\n\r\nExample:\
        \ \"Bearer 12345abcdef\""
      name: Authorization
      in: header